requirements.txt
fastapi
uvicorn
httpx
apscheduler
pydantic-settings
python-dotenv
scikit-learn
numpy
pandas
